rename : chat() 위치 변경 및 코드 정렬
